requirements.txt
django-embed-video
CKeditor
wysiwyg